requirements.txt
Flask>=2.2.0
pandas>=2.2.0
reportlab>=4.0.0
requests>=2.30.0
numpy>=1.22.0
gspread>=6.0.0
WooCommerce>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0